Add a Navigation Drawer following Material Design guidelines
The application is based on one activity, called DrawerActivity, which governs the logic of the Navigation
Drawer and it changes the content of this activity by fragments, depending on the selected option.

For now there are two small bugs:

* The status bar is darker than in other applications. For now it is a minor problem, at least in Marshmallow.
* The title of the Toolbar resizes in landscape. As I read, it seems a normal occurrence, but I do not like it,
so in the future, we will be playing stylize the Toolbar.

Signed-off-by: Álvaro Orduna León <alvaro.orduna.92@gmail.com>